Split out Entity, EntityQuerySet & get_word_count
